requirements.txt
weaviate-client==4.9.4
openai==1.55.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.0
pyyaml==6.0.1
loguru==0.7.2
python-json-logger==2.0.7
